+++ src/CHANGELOG.md
@@ -35,6 +35,7 @@
 - File Area: Update File Area
 - Utils: Update CLTK's Norse (Old) stop word list
 - Work Area: Rename Overview to Profiler
+- Work Area: Rename Wordlist to Wordlist Generator
 - Work Area: Update Work Area - Figure Settings - Sort by File
 
 ### ❌ Removals